third_party/blink/web_tests/fast/events/event-isolated-world-expected.txt

Tests that properties of various events do not leak between isolated worlds (bug 85158).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS documentObject is null
PASS document.pageDefinedVar is undefined.
PASS documentObject is null
PASS document.pageDefinedVar is undefined.
PASS documentObject is null
PASS document.pageDefinedVar is undefined.
PASS successfullyParsed is true

TEST COMPLETE
